nnoriter_Merge
Merge

+++ nnoriter-master/app/src/main/java/com/example/noriter/pcinfo_item.java
@@ -9,8 +9,8 @@ public class pcinfo_item {
     private String ad_rns ;
     private String ad_rn ;
     private String number ;
-    private int lo_x;
-    private int lo_y;
+    private double lo_x;
+    private double lo_y;
     private int CPU_B;
     private int CPU_G;
     private String CPU_C;
@@ -39,10 +39,10 @@ public class pcinfo_item {
     public void setNumber(String number1) {
         number = number1 ;
     }
-    public void setLo_x(int lo_x1) {
+    public void setLo_x(double lo_x1) {
         lo_x = lo_x1 ;
     }
-    public void setLo_y(int lo_y1) { lo_y = lo_y1 ; }
+    public void setLo_y(double lo_y1) { lo_y = lo_y1 ; }
     public void setCPU_B(int CPU1) { CPU_B = CPU1 ; }
     public void setCPU_G(int CPU1) { CPU_G = CPU1 ; }
     public void setCPU_C(String CPU1) { CPU_C = CPU1 ; }
@@ -61,8 +61,8 @@ public class pcinfo_item {
     public String getAd_rns() { return ad_rns; }
     public String getAd_rn() { return ad_rn; }
     public String getNumber() { return number; }
-    public int getLo_x() { return lo_x; }
-    public int getLo_y() { return lo_y; }
+    public double getLo_x() { return lo_x; }
+    public double getLo_y() { return lo_y; }
     public int getCPU_B() { return CPU_B; }
     public int getCPU_G() { return CPU_G; }
     public String getCPU_C() { return CPU_C; }
